requirements.txt
langchain
langchain-huggingface
langchain-pinecone
pinecone-client
sentence-transformers
langchain-community
flask
langchain-groq
langchain-core
python-dotenv
gunicorn
langchain-openai
firebase_admin